+++ previous-versions/accounting/src/sdk/sdk.ts
@@ -77,9 +77,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.39.0";
-    genVersion = "2.139.1";
-    userAgent = "speakeasy-sdk/typescript 0.39.0 2.139.1 3.0.0 @codat/accounting";
+    sdkVersion = "0.40.0";
+    genVersion = "2.142.2";
+    userAgent = "speakeasy-sdk/typescript 0.40.0 2.142.2 3.0.0 @codat/accounting";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);